lần 3

+++ src/main/java/model/ClientStat.java
@@ -17,12 +17,13 @@ public class ClientStat implements Serializable {
     private Client client;
 
     @Column(name = "day")
-    private int day;
+    private long day;
 
     @Column(name = "payment")
-    private float payment;
+    private double payment;
 
-    public ClientStat() {}
+    public ClientStat() {
+    }
 
     public int getClientStatId() {
         return clientStatId;
@@ -40,19 +41,19 @@ public class ClientStat implements Serializable {
         this.client = client;
     }
 
-    public int getDay() {
+    public long getDay() {
         return day;
     }
 
-    public void setDay(int day) {
+    public void setDay(long day) {
         this.day = day;
     }
 
-    public float getPayment() {
+    public double getPayment() {
         return payment;
     }
 
-    public void setPayment(float payment) {
+    public void setPayment(double payment) {
         this.payment = payment;
     }
 }
